requirements.txt
playwright~=1.55.0
python-dotenv~=1.1.1